+++ ext/VisualiserExt/VisualiserExt.jl
@@ -2,27 +2,65 @@
 
 module VisualiserExt
 
+
+############ Module dependencies ############
+
+import Base: @lock, @lock_nofail
+
 if isdefined(Base, :get_extension)
     using MuJoCo
     import MuJoCo.LibMuJoCo
-    import MuJoCo.Visualiser
-    import GLFW
+    import MuJoCo.LibMuJoCo: Model, Data, mjrRect, mjr_render
+    using MuJoCo.Visualiser
+
     using FFMPEG
+    using GLFW: GLFW, Window, Key, Action, MouseButton, GetKey, RELEASE, PRESS, REPEAT
+    using Observables: Observable, on, off
     using PrettyTables: pretty_table
     using Printf: @printf
     using StaticArrays
 else
     using ..MuJoCo
     import ..MuJoCo.LibMuJoCo
-    import ..MuJoCo.Visualiser
-    import ..GLFW
+    import ..MuJoCo.LibMuJoCo: Model, Data, mjrRect, mjr_render
+    using ..MuJoCo.Visualiser
+
     using ..FFMPEG
+    using ..GLFW: GLFW, Window, Key, Action, MouseButton, GetKey, RELEASE, PRESS, REPEAT
+    using ..Observables: Observable, on, off
     using ..PrettyTables: pretty_table
     using ..Printf: @printf
     using ..StaticArrays
 end
 
+
+############ Constants ############
+
+const Maybe{T} = Union{T, Nothing}  # From LyceumBase.jl
+const MAXGEOM = 10000               # preallocated geom array in mjvScene
+const MIN_REFRESHRATE = 30          # minimum rate when sim can't run at native refresh rate
+const RNDGAMMA = 0.9
+
+const RES_HD = (1280, 720)
+const RES_FHD = (1920, 1080)
+const RES_XGA = (1024, 768)
+const RES_SXGA = (1280, 1024)
+
+
+############ Includes ############
+
+include("util.jl")
+include("glfw.jl")
+include("ratetimer.jl")
+include("types.jl")
+include("functions.jl")
+include("modes.jl")
+include("defaulthandlers.jl")
+
 include("visualiser.jl")
+
+
+############ Functions ############
 
 function __init__()
     if Threads.nthreads() == 1
@@ -52,4 +90,4 @@ function MuJoCo.Visualiser.test_visualiser()
     visualise(model, data, controller=ctrl!)
 end
 
-end
+end # end VisualiserExt
